feature: added simplified inheritance model base
